+++ src/components/v1/cta/index.ts
@@ -1,0 +1,1 @@
+export * from './alpha-feedback';
