+++ src/components/v1/index.ts
@@ -1,3 +1,4 @@
+export * from './cta';
 export * from './icons';
 export * from './layouts';
 export * from './palette';
